Mertre root6 with upstream/v6-24-00-patches

+++ graf2d/asimage/src/libAfterImage/asim_afterbase.h
@@ -31,7 +31,6 @@
 /* from libAfterBase/astypes.h : */
 
 #include <stdio.h>
-#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -304,7 +303,7 @@ typedef union ASHashableValue
 }
 ASHashableValue;
 #else
-typedef uintptr_t ASHashableValue;
+typedef unsigned long ASHashableValue;
 #endif
 
 typedef union ASHashData
@@ -324,7 +323,7 @@ typedef union ASHashData
 	CARD8  c8 ;
 }ASHashData;
 
-#define AS_HASHABLE(v)  ((ASHashableValue)((uintptr_t)(v)))
+#define AS_HASHABLE(v)  ((ASHashableValue)((unsigned long)(v)))
 
 typedef struct ASHashItem
 {
