+++ math/minuit2/inc/Minuit2/SqrtLowParameterTransformation.h
@@ -1,5 +1,5 @@
 // @(#)root/minuit2:$Id$
-// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei, E.G.P. Bos   2003-2017
+// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei   2003-2005
 
 /**********************************************************************
  *                                                                    *
@@ -35,13 +35,13 @@ public:
    ~SqrtLowParameterTransformation() {}
 
    // transformation from internal to external
-   long double Int2ext(long double Value, long double Lower) const;
+   double Int2ext(double Value, double Lower) const;
 
    // transformation from external to internal
-   long double Ext2int(long double Value, long double Lower, const MnMachinePrecision &) const;
+   double Ext2int(double Value, double Lower, const MnMachinePrecision &) const;
 
    // derivative of transformation from internal to external
-   long double DInt2Ext(long double Value, long double Lower) const;
+   double DInt2Ext(double Value, double Lower) const;
 
 private:
 };
