+++ math/minuit2/inc/Minuit2/SqrtUpParameterTransformation.h
@@ -1,5 +1,5 @@
 // @(#)root/minuit2:$Id$
-// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei, E.G.P. Bos   2003-2017
+// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei   2003-2005
 
 /**********************************************************************
  *                                                                    *
@@ -36,13 +36,13 @@ public:
    ~SqrtUpParameterTransformation() {}
 
    // transformation from internal to external
-   long double Int2ext(long double Value, long double Upper) const;
+   double Int2ext(double Value, double Upper) const;
 
    // transformation from external to internal
-   long double Ext2int(long double Value, long double Upper, const MnMachinePrecision &) const;
+   double Ext2int(double Value, double Upper, const MnMachinePrecision &) const;
 
    // derivative of transformation from internal to external
-   long double DInt2Ext(long double Value, long double Upper) const;
+   double DInt2Ext(double Value, double Upper) const;
 
 private:
 };
