+++ math/minuit2/src/SqrtLowParameterTransformation.cxx
@@ -1,5 +1,5 @@
 // @(#)root/minuit2:$Id$
-// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei, E.G.P. Bos   2003-2017
+// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei   2003-2005
 
 /**********************************************************************
  *                                                                    *
@@ -19,29 +19,28 @@ namespace ROOT {
 
 namespace Minuit2 {
 
-long double SqrtLowParameterTransformation::Int2ext(long double value, long double lower) const
+double SqrtLowParameterTransformation::Int2ext(double value, double lower) const
 {
    /// internal to external transformation
-   long double val = lower - 1. + std::sqrt(value * value + 1.);
+   double val = lower - 1. + std::sqrt(value * value + 1.);
    return val;
 }
 
-long double
-SqrtLowParameterTransformation::Ext2int(long double value, long double lower, const MnMachinePrecision &) const
+double SqrtLowParameterTransformation::Ext2int(double value, double lower, const MnMachinePrecision &) const
 {
    // external to internal transformation
-   long double yy = value - lower + 1.;
-   long double yy2 = yy * yy;
+   double yy = value - lower + 1.;
+   double yy2 = yy * yy;
    if (yy2 < 1.)
       return 0;
    else
       return std::sqrt(yy2 - 1);
 }
 
-long double SqrtLowParameterTransformation::DInt2Ext(long double value, long double) const
+double SqrtLowParameterTransformation::DInt2Ext(double value, double) const
 {
    // derivative of internal to external transofrmation   :  d (Int2Ext) / d Int
-   long double val = value / (std::sqrt(value * value + 1.));
+   double val = value / (std::sqrt(value * value + 1.));
    return val;
 }
